+++ nix/magic.nix
@@ -11,6 +11,27 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+# Copyright (c) 2003-2023 Eelco Dolstra and the Nixpkgs/NixOS contributors
+
+# Permission is hereby granted, free of charge, to any person obtaining
+# a copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish,
+# distribute, sublicense, and/or sell copies of the Software, and to
+# permit persons to whom the Software is furnished to do so, subject to
+# the following conditions:
+
+# The above copyright notice and this permission notice shall be
+# included in all copies or substantial portions of the Software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+# NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+# LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+# WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 {
   pkgs ? import ./pkgs.nix,
   rev,
@@ -19,7 +40,8 @@
   python-pname ? "python38Full",
 }:
 
-with pkgs; magic-vlsi.overrideAttrs (finalAttrs: previousAttrs: {
+with pkgs; stdenv.mkDerivation rec {
+  pname = "magic-vlsi";
   version = "${rev}";
 
   src = fetchgit {
@@ -29,6 +51,28 @@ with pkgs; magic-vlsi.overrideAttrs (finalAttrs: previousAttrs: {
   };
 
   nativeBuildInputs = [ pkgs."${python-pname}" tcsh gnused ];
+
+  # So here's the situation:
+  ## * Magic will not compile without X11 libraries, even for headless use.
+  ## * Magic with Cairo will use Cairo-X11 headers that are not available on
+  ## pre-built Mac versions of Cairo which presume you're going to use Quartz.
+  ## -> Unintuitively, that means X11 libraries must be present on Mac, but not
+  ## Cairo.
+  buildInputs = [
+    xorg.libX11
+    m4
+    mesa_glu
+    ncurses
+    tcl
+    tcsh
+    tk
+  ] ++ lib.optionals stdenv.isLinux [cairo];
+
+  configureFlags = [
+    "--with-tcl=${tcl}"
+    "--with-tk=${tk}"
+    "--disable-werror"
+  ];
 
   preConfigure = ''
   # Replace csh invocations with tcsh
@@ -45,4 +89,9 @@ with pkgs; magic-vlsi.overrideAttrs (finalAttrs: previousAttrs: {
   # Patch all shebangs to meet Nix standards
   patchShebangs ./scripts/*
   '';
-})
+
+  NIX_CFLAGS_COMPILE = "-Wno-implicit-function-declaration -Wno-parentheses -Wno-macro-redefined";
+
+  # Fairly sure this is deprecated?
+  enableParallelBuilding = true;
+}