+++ openlane/flows/misc.py
@@ -23,7 +23,6 @@ from .flow import Flow
 from ..common import get_script_dir
 from ..steps import Step, State, KLayout
 
-
 @Flow.factory.register()
 class OpenInKLayout(Flow):
     """